fix icon state
The problem was that the isNavbarVisible state, which is responsible for displaying/hiding the menu, was in the Header component. When you clicked on a link inside the menu, the route would be updated, the Header would be rendered, and the isNavbarVisible state would be set back to its original value (false), hiding the menu.